add_confirmed

+++ src/pages/Home.js
@@ -5,6 +5,7 @@ import { SideMenu } from "../components/SideMenu";
 import { postRepository } from "../repositories/Post";
 import { Post } from "../components/Post";
 import { Pagination } from "../components/Pagination";
+import { authRepository } from "../repositories/auth";
 
 const limit = 5;
 
@@ -12,7 +13,7 @@ function Home() {
     const[content, setContent] = useState('');
     const[posts,setPosts] = useState([])
     const[page,setPage] = useState(1)
-    const {currentUser} = useContext(SessionContext);
+    const {currentUser, setCurrentUser} = useContext(SessionContext);
 
     useEffect(() => {
         fetchPosts();
@@ -47,9 +48,27 @@ function Home() {
     }
 
     const deletePost = async (postId) => {
+        const confirmed = window.confirm("本当に削除しますか？");
+
+        //確認ダイアログ
+        if (!confirmed) {
+            return; 
+        }
         await postRepository.delete(postId);
         setPosts(posts.filter((post) => post.id !== postId));
     };
+
+    const signout = async () => {
+        const confirmed = window.confirm("本当にログアウトしますか？");
+
+        //確認ダイアログ
+        if (!confirmed) {
+            return; 
+        }
+    
+        await authRepository.signout();
+        setCurrentUser(null);
+    }
     
     if(currentUser == null) return <Navigate replace to="/signin"/>;
 
@@ -58,7 +77,9 @@ function Home() {
         <header className="bg-[#34D399] p-4">
             <div className="container mx-auto flex items-center justify-between">
             <h1 className="text-3xl font-bold text-white">SNS APP</h1>
-            <button className="text-white hover:text-red-600">ログアウト</button>
+            <button
+            onClick={signout}
+            className="text-white hover:text-red-600">ログアウト</button>
             </div>
         </header>
         <div className="container mx-auto mt-6 p-4">
